Add a new amt_urlview.py command
Alternative implementation to the "urlview" command.  The main benefit
is improved configurability for choosing which URLs to use, and
optionally demangling URLs before displaying them.

The user config file can provide a function to guess the preferred URL
to display.

This command is still a little bit rough around the edges at the moment,
but functional for basic usage.