[imap] fix file descriptor leak on connect failure
Add a __del__() method to ConnectionCore, to ensure that the interrupt
FDs are closed even if close() is never explicitly called on the
ConnectionCore object.  This was previously causing an fd leak if the
imap.Connection constructor threw an exception trying to connect to the
server.

Also move the logic to create the interrupt pipe until after we
successfully connect to the server.  This way we don't even create the
pipe FDs until we have a successful connection.