[prune] revert to only issuing 256 deletes at once
Current versions of Exchange on office365.com appear to simply close the
connection on receipt of a STORE request with a large number of message
IDs.  I didn't play around too closely with the limits, but 1024
message IDs at once appears to be too large.  Using a limit of 256
appears to work reliably.